eksperimen: sistem sliding window dengan look-ahead buffer

+++ firmware_arduino/lib/PengolahSinyalPPG/src/PengolahSinyalPPG.cpp
@@ -1,7 +1,7 @@
 #include "PengolahSinyalPPG.h"
 
 // Fungsi mencari nilai tertinggi di buffer untuk penentuan threshold dinamis
-float cariNilaiMax(float *sinyalBuffer, uint8_t panjangBuffer) {
+float cariNilaiMax(uint16_t *sinyalBuffer, uint8_t panjangBuffer) {
   float nilaiMax = sinyalBuffer[0];
   for (uint8_t i = 1; i < panjangBuffer; i++) {
     if (sinyalBuffer[i] > nilaiMax) {
